# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Programowanie usług serwerowych",
    "description": "API Programowanie usług serwerowych",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/project/{project_id}/task": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "priority": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "estimated_end_date": {
                  "type": "string",
                  "format": "date"
                }
              },
              "required": [
                "name",
                "description",
                "priority",
                "estimated_end_date"
              ]
            }
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                },
                "_links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ref": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/{project_id}/task/{task_id}?": {
      "get": {
        "parameters": [
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "task_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "project_id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "priority": {
                            "type": "number"
                          },
                          "estimated_end_date": {
                            "type": "string",
                            "format": "date"
                          }
                        }
                      }
                    }
                  }
                },
                "_links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ref": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "head": {
        "parameters": [
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "task_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "project_id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "priority": {
                            "type": "number"
                          },
                          "estimated_end_date": {
                            "type": "string",
                            "format": "date"
                          }
                        }
                      }
                    }
                  }
                },
                "_links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ref": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/{project_id}/task/{task_id}": {
      "put": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "priority": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "estimated_end_date": {
                  "type": "string",
                  "format": "date"
                }
              }
            }
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "task_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {}
                },
                "_links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ref": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "task_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {}
                },
                "_links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ref": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/{project_id}/user/add": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "number"
                }
              },
              "required": [
                "user_id"
              ]
            }
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/project/{project_id}/user/list": {
      "get": {
        "parameters": [
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "head": {
        "parameters": [
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/project/{project_id}/user/delete": {
      "delete": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "number"
                }
              },
              "required": [
                "user_id"
              ]
            }
          },
          {
            "type": "number",
            "required": true,
            "in": "path",
            "name": "project_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}